+++ src/main/java/bookers/bookkeeper/userprofile/UserProfile.java
@@ -6,6 +6,7 @@ import bookers.bookkeeper.user.User;
 import lombok.Data;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Entity
@@ -19,7 +20,7 @@ public class UserProfile {
     @OneToOne
     User user;
 
-    Float meanScore;
+    Double meanScore;
 
     Integer pagesRead;
