+++ src/main/java/bookers/bookkeeper/userprofile/dto/UserProfileDTO.java
@@ -4,12 +4,13 @@ import bookers.bookkeeper.author.dto.AuthorDTO;
 import bookers.bookkeeper.book.dto.BookDTO;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 public class UserProfileDTO {
     private String userName;
-    private Float meanScore;
+    private Double meanScore;
     private Integer pagesRead;
     private Integer booksRead;
     private List<BookDTO> favoriteBooks;
